Add query map and module for queries (#62)
* Add query map and module for queries

* Remove extraneous file

* indentation fixes and imports added
