+++ listenbrainz_spark/query_map.py
@@ -1,0 +1,5 @@
+from listenbrainz_spark import stats
+
+functions = {
+	'stats.user.all': stats.user.all.calculate,
+}
